apicreditodebits/requirements.txt
Django==2.2.3
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
Markdown==3.1.1
psycopg2==2.7.7
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
